Ceph hostname change during upgrade
L>M has changed the hostname of our containers, therefore we
need to adapt this everywhere. Ceph was left behind, and this
is fixing the hostnames by shutting down the monitors, removing the mons,
deleting the containers, recreating them, and rejoining the cluster, all
of this in a serial way.

The ceph-mon playbook was adapted to verify the bind mount is there, in
order to avoid data loss.

Signed-off-by: Jean-Philippe Evrard <jean-philippe.evrard@rackspace.co.uk>